requirements.txt
streamlit
mecab-python3
wordcloud
pandas
ipadic